Make spotlight placement independent of box position and size
Box used to be centered in top left corner, with a fixed size.
The mouse position code relied on this position and size in
calculating the placement of the spotlight.

New code is independent of the position and size of the box (it
uses relative offsets and element sizes provided by jQuery).

As an added bonus, it uses jQuery mouse event's e.pageX and pageY
values, which jQuery explicitly standardizes, as opposed to
e.clientX and clientY, which may vary across browsers/platforms.