Prevent multiple firings of events on start and target
Before, when the start button was pressed, and sometimes when the
target was reached, this would fire off multiple events. These
would launch multiple separate trials and create lots of other
problems like that.

To fix this, we keep a count ("semaphore") and only allow actions
to take place if this is the first time it's been fired.