Display average time to target to users
Rather than displaying the current round number or the time
elapsed in the previous round, display their mean time to target
(over all rounds).

Hopefully, this enhances the game-like aspects of this setup and
motivates the users to keep looking for targets (and giving us
more data). Hey, it worked on me! ;-)